requirement.txt
numpy
scikit-learn
tensorflow
matplotlib
Pillow
tqdm
keras
seaborn